test(NSC): reorganize test files per reviewer feedback
Move tests from hardening_test.go into the files requested by aauren:
- TestShuffleDoesNotPanicOnEmptySlice → network_services_controller_test.go
- TestSetupMangleTableRuleRejectsInvalidIP → network_services_controller_test.go
- TestNodePortHealthCheckConcurrentAccess → new nodeport_healthcheck_test.go

Delete hardening_test.go (now empty).
